+++ initglobals.py
@@ -1,0 +1,4 @@
+counter = 0
+trigger = False
+oldcounter = -1
+activemenu = 0 #defaults to IDLE screen
